Updated CRAM to v3.1.
The description of rANS has been removed as it is now in the
CRAMcodecs.tex document.  Added brief paragraphs for the other new
v3.1 codecs.

Updated the Block Structure table to indicate the internal code
numbers for the newer codecs and also added version information so it
is clear in which version they first appeared.

The intention here is we should be able to read the CRAM v3.1 document
to understand how to create a version 3.0 format file (as the sole
differences so far are the addition of new compression codecs).