Clarify the multiple incompatible uses of "phred-scale".
Sometimes this refers to $10 log_{10}(p)$, sometimes to $10
log_{10}(1-p)$, and sometimes to something normalised so $p$ isn't
really a probability at all.

Note CNL, CNP and CNQ don't mention phred anywhere in their
short description and only Phred in the long description for CNQ, so
I applied the same logic to PL, PP (is this correct?) and PQ.

Also clarified the "VCF tag naming conventions" part.  I changed
phred-scale in one part there to phred-true-scale.  I'm not so happy
with that, but as it's immediately followed by the formula I think
it's clear.